Move Traceable, trace_reflector from utils.rs to trace.rs.(fixes #1748)
